requirements/base.txt
boto3==1.12.41
botocore==1.15.41
distro==1.5.0
Django==2.2.12
django-environ==0.4.5
django-storages==1.9.1
djangorestframework==3.11.0
docutils==0.15.2
jmespath==0.9.5
numpy==1.18.2
pandas==0.24.2
pkg-resources==0.0.0
psycopg2-binary==2.8.5
python-dateutil==2.8.1
pytz==2019.3
s3transfer==0.3.3
six==1.14.0
sqlparse==0.3.1
tabula-py==1.4.3
urllib3==1.25.9
